extended HWUart with pseudoterminal
Added a HWUart subclass that is attached to a pseudoterminal, so other
processes can communicate to the simulation as if it were connected via
serial.  The pty bypasses bit-encoding/decoding but maintains the timing,
and supports up to 1M baud, which is convenient for co-operating processes.

One mystery: I found that occasional tx chars were repeated, and
delaying setting UDRE fixed the problem, but I didn't investigate
further.